OpenSSL 3 fixes: Fix "dirty shutdown" detection for OpenSSL 3
OpenSSL 3.0 introduces SSL_OP_IGNORE_UNEXPECTED_EOF to "ignore" a remote
not sending a close notify before closing the connection. We don't want
to enable that (and we want to be able to distinguish the two cases in
tests), but even when not enabled, this new functionality changes the
error codes produced when reading from an SSL bio, such that we must
also check for SSL_R_UNEXPECTED_EOF_WHILE_READING as well as the raw
system call error.